+++ test/core/optimize.jl
@@ -1,0 +1,5 @@
+using Test
+using RecurrenceMicrostatesAnalysis
+
+@test_throws ArgumentError optimize(Threshold(), RecurrenceEntropy())
+@test_throws ArgumentError optimize(Threshold(), Disorder())