+++ test/recurrences/recurrences.jl
@@ -1,0 +1,2 @@
+testfile("corridor.jl")
+testfile("standard.jl")