+++ test/rqa/rr.jl
@@ -1,0 +1,12 @@
+using Test
+using Distributions
+using RecurrenceMicrostatesAnalysis
+
+x = StateSpaceSet(rand(1000))
+dist = distribution(x, 0.27, 4)
+
+@test 0 ≤ measure(RecurrenceRate(), dist) ≤ 1
+@test 0 ≤ measure(RecurrenceRate(), x) ≤ 1
+
+@test measure(RecurrenceRate(), dist) isa Real
+@test measure(RecurrenceRate(), x) isa Real