+++ test/sampling/sampling.jl
@@ -1,0 +1,2 @@
+testfile("full.jl")
+testfile("random.jl")