+++ test/shapes/shapes.jl
@@ -1,0 +1,3 @@
+testfile("diagonal.jl")
+testfile("rect.jl")
+testfile("triangle.jl")